clone the original conv_next to repositories in launch.py, add convnext requirements to requirements.txt
